Remove default audience from GitLab
GitLab claims that there is a default audience (domain of the instance)
but there is no way to actually access that since the pipeline yaml
requires an audience: make sure we don't try using the default audience.

This changes tests a bit because it happened to be the case we used for
detect_credentials test case.